add version of Boost libraries to readme
The given minimum version for Boost is 1.69, but it may be that
even older versions may work. What is known is that versions 1.71
and 1.69 do work, but version 1.66 does not work. So it is
currently not clear, whether versions 1.67 and 1.68 do work or
not. Those versions have not been tested.