gandalf_the_grey: fix potentially wrong size for file names on Windows
The wide-character string that is passed to the path constructor
has potentially several null characters at the end. In practice
that should not be a problem, because C-like OS APIs stop reading
strings at the first null character. But let's not take any
chances here.

Linux builds are not affected by this problem.